Split shared/plant_contract.py into one module per contract layer
plant_contract.py was 1,822 lines. Its own docstring already described
four separable concerns, so make that structure real: one submodule per
layer, over a shared base of digesting and version loading, with manifest
generation and enforcement on top.

  constants.py       paths + schema ids
  errors.py          the two exceptions
  digests.py         portable canonicalisation and semantic digests
  identity.py        PlantVersion / PlantIdentity
  versions.py        reviewed revisions and committed manifest, from disk
  introspection.py   shared MuJoCo model-introspection helpers
  source_layer.py    MJCF and referenced asset bytes
  policy_layer.py    observations, actions, sensors, controls
  physics_layer.py   dynamics, contacts, actuators, reset state
  visual_layer.py    render geometry, materials, cameras, lights
  manifest.py        build / render / write / check
  validation.py      fail-closed checks for models, envs, checkpoints
  __main__.py        the --check/--write/--show CLI

The package __init__ re-exports the same public surface, so no caller
changes and `python -m environments.shared.plant_contract --check` keeps
working. _load_generated_manifest moves down into versions.py, since
_species_entries and load_plant_versions fall back to it — leaving it in
manifest.py would have made the layering circular.

The configs/ path constants are now read through the constants module
rather than bound by value at import, which gives the tests a single
patch point; test_plant_contract.py patches that module and imports the
private helpers it exercises from the submodule that owns them.
